feat(online-evals): move the evaluator status badge into the Activity panel
